services/desktop_gui_frontend/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
mypy==1.4.0
mypy-extensions==1.0.0
numpy==1.25.0
packaging==23.1
Pillow==10.0.0
pydantic==1.10.9
pyparsing==3.0.9
pyproject_hooks==1.0.0
PySimpleGUI==4.60.5
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
types-python-dateutil==2.8.19.13
types-requests==2.31.0.2
types-urllib3==1.26.25.13
typing_extensions==4.6.3
urllib3==2.0.3
